requirements.txt
colbert-ai==0.2.19
faiss-cpu==1.7.4
torch>=2.1.0
transformers>=4.39.0
openai>=1.37.0
python-docx>=0.8.11
pypdf>=4.2.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
